Fix #1 Making Modular Code
