+++ choices.rb
@@ -1,0 +1,15 @@
+module Choices
+  PREORDER = 1
+  POSTORDER = 2
+  INORDER = 3
+  LEVELORDER = 4
+  LARGEST = 5
+  SMALLEST = 6
+  ADD_ELEMENT = 7
+  SEARCH_ELEMENT = 8
+  REMOVE_ELEMENT = 9
+  PATH = 10
+  OPEN_FILE = 11  
+
+         
+end
