+++ tree.rb
@@ -1,0 +1,10 @@
+#Tree Structure
+class Tree
+  attr_accessor :val, :left, :right
+
+  def initialize(val)
+    @val=val
+    @left=nil
+    @right=nil
+  end
+end
